efm32/tempdrv: try to add the driver for furture
TEMPDRV gives the user a nice interface to the EMU internal temperature
sensor which is present on the EFR32 and some EFM32 devices. The TEMPDRV
supports application specific callbacks at given temperature thresholds.
The EMU internal temperature sensor is running in EM0-EM4H and is capable
of waking up the core on temperature change. The EMU temperature sensor is
running continuously and measurements are taken every 250 ms.

The TEMPDRV is using the EMU peripheral and not the ADC peripheral. The
ADC contains another internal temperature sensor which is not touched by
the TEMPDRV.

This low power device only available in following mcu:

	EFM32JG12B
	EFM32JG1B
	EFM32PG12B
	EFM32PG1B

	EFR32BG12P
	EFR32BG13P
	EFR32BG1B
	EFR32BG1P
	EFR32BG1V
	......